Make first and last name required
- Decided I want to force users to give me first and last name at least
so I can display them in lists of users and such, so the "Edit later"
link only appears for users who have already given first and last name
(this was updated in the view with the Profile#required_info? method)
- Date of birth and bio are now optional but the user will be prompted
to update them every time they log in (still to be implemented).
- Profile#age runs a check for the existence of a date of birth before
doing anything else