Remove "rejected" status from Friendship enum
- Rejected requests will be destroyed from the database. This change
happens because I was considering the implications of having a
friendship with a "rejected" status on the database. One of the
difficulties would be if the request was then to be sent from the other
direction (from the rejecter). In this situation I would have a
"requestee" sending a request and even though it could be made to work
there is no need to go this route. If I later want a rejected status
(useful maybe to keep track of rejection count) I might separate
friendships from requests