Create stimulus controller for confirmations
- I had a problem where upon deleting a post in its show page due
to turbo streams the user would stay in the same page with no relevant
info. I felt this wasn't desired behaviour and that the user should
be redirected back to the index page. For this I had to disable turbo
which made the confirm dialog box stop working, so I had to create a
stimulus controller to simulate the turbo-confirm behaviour.
- I now need to have a conditional so that the delete button works
differently in the index and in the show page